backend/requirements.txt
pycantonese
flask
flask-cors
tencentcloud-sdk-python